Resolved the indexing issue for most recent strategy (reverse while loop)
All this took is modifying the message.py-->parseStateChange so that
the 0th column does indeed hold the scores (note: The ordering of numbers
in each row matches that of the CHANGE; message). Furthermore now bestPit
in makeMove() points to the correct index and does not need to be incremented.